added functionalities to apply workedhours, based on that data will first be schedules made to provide for the worked hours.
added BID column to employee, since that's easier to search for employees from the given BID in the csv file

known issues: need a check to remove special charachters or allow special characters when inserting a name like noëlle since the ë isnt supported. this causes some issues with inserting employees.

Tested this with issue in add employee and cant add a employee named Noëlle